requirements.txt
python-binance==1.0.13
websockets==8.1
pandas==1.4.2